Refactoring of the MD components.
