+++ src/core/ui/forms/MarkdownInput/MarkdownInput.tsx
@@ -1,35 +1,15 @@
-import React, {
-  FormEvent,
-  memo,
-  useCallback,
-  useEffect,
-  useImperativeHandle,
-  useLayoutEffect,
-  useRef,
-  useState,
-  useMemo,
-} from 'react';
+import React, { memo, useCallback, useEffect, useImperativeHandle, useRef } from 'react';
 import { Box, useTheme } from '@mui/material';
-import { useTranslation } from 'react-i18next';
-import { Editor, EditorContent, useEditor } from '@tiptap/react';
-import StarterKit from '@tiptap/starter-kit';
 import { InputBaseComponentProps } from '@mui/material/InputBase/InputBase';
-import { CharacterCountContainer, useSetCharacterCount } from './CharacterCountContext';
+import { CharacterCountContainer } from './CharacterCountContext';
 import MarkdownInputControls from '../MarkdownInputControls/MarkdownInputControls';
-import { Image } from '@tiptap/extension-image';
-import { Link } from '@tiptap/extension-link';
-import usePersistentValue from '@/core/utils/usePersistentValue';
-import UnifiedConverter from '@/core/ui/markdown/html/UnifiedConverter';
 import { gutters } from '@/core/ui/grid/utils';
-import { EditorState } from '@tiptap/pm/state';
-import { Highlight } from '@tiptap/extension-highlight';
-import { Selection } from 'prosemirror-state';
-import { EditorOptions } from '@tiptap/core';
-import { Iframe } from '../MarkdownInputControls/InsertEmbedCodeButton/Iframe';
-import { EditorView } from '@tiptap/pm/view';
-import { useUploadFileMutation } from '@/core/apollo/generated/apollo-hooks';
-import { useNotification } from '../../notifications/useNotification';
+import { EditorContent, Editor } from '@tiptap/react';
 import { useStorageConfigContext } from '@/domain/storage/StorageBucket/StorageConfigContext';
+import { useEditorConfig } from './hooks/useEditorConfig';
+import { useImageUpload } from './hooks/useImageUpload';
+import { useMarkdownInputUI } from './hooks/useMarkdownInputUI';
+import { useMarkdownEditor } from './hooks/useMarkdownEditor';
 
 export interface MarkdownInputProps extends InputBaseComponentProps {
   controlsVisible?: 'always' | 'focused';
@@ -48,8 +28,6 @@ export interface MarkdownInputRefApi {
   value: string | undefined;
   getLabelOffset: () => Offset;
 }
-
-const ImageExtension = Image.configure({ inline: true });
 
 const proseMirrorStyles = {
   outline: 'none',
@@ -76,155 +54,49 @@ export const MarkdownInput = memo(
   }) => {
     const containerRef = useRef<HTMLDivElement>(null);
     const toolbarRef = useRef<HTMLDivElement>(null);
-
-    const [hasFocus, setHasFocus] = useState(false);
-    const [isControlsDialogOpen, setIsControlsDialogOpen] = useState(false);
-    const isInteractingWithInput = hasFocus || isControlsDialogOpen;
-
-    const [htmlContent, setHtmlContent] = useState('');
-
-    const { markdownToHTML, HTMLToMarkdown } = usePersistentValue(UnifiedConverter());
-
-    const storageConfig = useStorageConfigContext();
-
-    const updateHtmlContent = async () => {
-      const content = await markdownToHTML(value);
-      setHtmlContent(String(content));
-    };
-
-    const { t } = useTranslation();
-
-    const notify = useNotification();
-
-    const [uploadFile] = useUploadFileMutation({
-      onCompleted: data => {
-        notify(t('components.file-upload.file-upload-success'), 'success');
-
-        editor?.commands.setImage({ src: data.uploadFileOnStorageBucket, alt: 'pasted-image' });
-      },
-
-      onError: error => {
-        console.error('File upload failed:', error.message);
-        notify(t('components.file-upload.file-upload-error'), 'error');
-      },
-    });
-
-    const isImageOrHtmlWithImage = (item: DataTransferItem, clipboardData: DataTransfer | null) => {
-      if (item.type.startsWith('image/') || (item.kind === 'file' && item.type.startsWith('image/'))) {
-        return true; // Image
-      }
-
-      if (item.kind === 'string' && item.type === 'text/html') {
-        const htmlContent = clipboardData?.getData('text/html');
-        return htmlContent?.includes('<img') ?? false; // HTML tag with image
-      }
-
-      return false; // Not an image or HTML with images
-    };
-
-    const storageBucketId = storageConfig?.storageBucketId;
-
-    /**
-     * Handles the paste event in the editor.
-     *
-     * @param _view - The editor view instance.
-     * @param event - The clipboard event triggered by pasting.
-     * @returns {boolean} - Returns true if the paste event is handled, otherwise false - continue execution of the default.
-     *
-     * Reference to alternative way of handling paste events in Tiptap: https://tiptap.dev/docs/editor/extensions/functionality/filehandler
-     *
-     */
-    const handlePaste = useCallback(
-      (_view: EditorView, event: ClipboardEvent): boolean => {
-        if (!storageBucketId) {
-          return false; // Allow default behavior for text
-        }
-
-        const clipboardData = event.clipboardData;
-        const items = clipboardData?.items;
-
-        if (!items) {
-          return false; // Allow default behavior for text
-        }
-
-        const itemsArray = Array.from(items); // Keep `Array.from` since if any kind of `for` loop is used it will iterate only over one item.
-
-        itemsArray.forEach(item => {
-          const isImage = isImageOrHtmlWithImage(item, clipboardData);
-
-          if (hideImageOptions && isImage) {
-            event.preventDefault();
-            return true; // Block paste of images or HTML with images
-          }
-
-          if (isImage) {
-            const file = item.getAsFile();
-
-            if (file) {
-              const reader = new FileReader();
-
-              reader.onload = () => {
-                uploadFile({ variables: { file, uploadData: { storageBucketId, temporaryLocation } } });
-              };
-
-              reader.readAsDataURL(file);
-              event.preventDefault();
-              return true; // Block default behavior for images
-            }
-          }
-        });
-
-        return false; // Allow default behavior for text
-      },
-      [storageBucketId, hideImageOptions, temporaryLocation, uploadFile, isImageOrHtmlWithImage]
-    );
-
-    const editorOptions: Partial<EditorOptions> = useMemo(
-      () => ({
-        extensions: [StarterKit, ImageExtension, Link, Highlight, Iframe],
-        editorProps: { handlePaste },
-      }),
-      [handlePaste]
-    );
-
-    const editor = useEditor({ ...editorOptions, content: htmlContent }, [htmlContent]);
-
-    if (!editor) {
-      return null;
-    }
-
-    // Currently used to highlight overflow but can be reused for other similar features as well
-    const shadowEditor = useEditor({ ...editorOptions, content: '', editable: false });
-
-    useLayoutEffect(() => {
-      if (!editor || !isInteractingWithInput || editor.getText() === '') {
-        updateHtmlContent();
-      }
-    }, [value, hasFocus]);
-
     const theme = useTheme();
 
-    const areControlsVisible = () => {
-      if (controlsVisible === 'always') {
-        return true;
-      }
-      if (controlsVisible === 'focused') {
-        return isInteractingWithInput;
-      }
-    };
+    const storageConfig = useStorageConfigContext();
+    const storageBucketId = storageConfig?.storageBucketId;
 
-    const getLabelOffset = () => {
-      const offsetY = areControlsVisible()
-        ? toolbarRef.current
-          ? `${toolbarRef.current.clientHeight + 20}px`
-          : gutters(3)(theme)
-        : gutters(1)(theme);
+    const {
+      areControlsVisible,
+      getLabelOffset,
+      handleFocus,
+      handleBlur,
+      handleDialogOpen,
+      handleDialogClose,
+      prevEditorHeight,
+      setPrevEditorHeight,
+      isInteractingWithInput,
+    } = useMarkdownInputUI({
+      controlsVisible,
+      disabled: false,
+      toolbarRef,
+      containerRef,
+      onFocus,
+      onBlur,
+    });
 
-      return {
-        x: gutters()(theme),
-        y: offsetY,
-      };
-    };
+    const { handlePaste } = useImageUpload({
+      storageBucketId,
+      hideImageOptions,
+      temporaryLocation,
+      editor: null, // Will be set after editor creation
+    });
+
+    const editorConfig = useEditorConfig({
+      handlePaste,
+      disabled: false,
+    });
+
+    const { editor, shadowEditor } = useMarkdownEditor({
+      value,
+      onChange,
+      maxLength,
+      editorConfig,
+      isInteractingWithInput,
+    });
 
     useImperativeHandle(
       ref,
@@ -235,139 +107,47 @@ export const MarkdownInput = memo(
           return editor?.getText();
         },
       }),
-      [editor, areControlsVisible()]
+      [editor, getLabelOffset]
     );
 
-    const setCharacterCount = useSetCharacterCount();
+    const keepScrollPositionOnEditorReset = useCallback(
+      (editorInstance: Editor) => {
+        const handleCreate = () => {
+          setPrevEditorHeight(0);
+        };
 
-    useLayoutEffect(() => {
-      setCharacterCount(editor?.getText().length ?? 0);
-    }, [editor]);
+        editorInstance.on('create', handleCreate);
 
-    const wrapIframeWithStyledDiv = (markdown: string): string =>
-      markdown.replace(
-        /<iframe[^>]*><\/iframe>/g,
-        iframe =>
-          `<div style='position: relative; padding-bottom: 56.25%; width: 100%; overflow: hidden; border-radius: 8px; margin-bottom: 10px;'>${iframe}</div>`
-      );
-
-    const emitChangeOnEditorUpdate = (editor: Editor) => {
-      const handleStateChange = async () => {
-        let markdown = await HTMLToMarkdown(editor.getHTML());
-
-        markdown = wrapIframeWithStyledDiv(markdown);
-
-        setCharacterCount(editor.getText().length);
-
-        onChange?.({
-          currentTarget: {
-            value: markdown,
-          },
-          target: {
-            value: markdown,
-          },
-        } as unknown as FormEvent<HTMLInputElement>);
-      };
-
-      editor.on('update', handleStateChange);
-
-      return () => {
-        editor.off('update', handleStateChange);
-      };
-    };
-
-    useEffect(() => {
-      if (editor) {
-        return emitChangeOnEditorUpdate(editor);
-      }
-    }, [editor]);
-
-    const updateShadowEditor = (editor: Editor, maxLength: number) => {
-      const highlightOverflow = () => {
-        if (!shadowEditor) {
-          return;
-        }
-
-        const contentLength = editor.getText().length;
-
-        if (contentLength <= maxLength) {
-          return;
-        }
-
-        try {
-          shadowEditor.view.updateState(EditorState.create({ doc: editor.state.doc }));
-        } catch (error) {
-          // In some states the "shadow" editor fails to update, but this doesn't break the highlight
-          console.error('Failed to update shadow editor state:', error);
-        }
-
-        const end = Selection.atEnd(shadowEditor.state.doc).from;
-        const overflow = contentLength - maxLength;
-
-        shadowEditor
-          .chain()
-          .setTextSelection({
-            from: end - overflow,
-            to: end,
-          })
-          .setHighlight() // Passing a highlight color doesn't work well here, styled later in sx={mark:{...}}
-          .run();
-      };
-
-      highlightOverflow();
-
-      editor.on('update', highlightOverflow);
-
-      return () => {
-        editor.off('update', highlightOverflow);
-      };
-    };
-
-    useEffect(() => {
-      if (editor && typeof maxLength === 'number') {
-        return updateShadowEditor(editor, maxLength);
-      }
-    }, [editor, maxLength]);
-
-    const [prevEditorHeight, setPrevEditorHeight] = useState(0);
-
-    const keepScrollPositionOnEditorReset = (editor: Editor) => {
-      const handleCreate = () => {
-        setPrevEditorHeight(0);
-      };
-
-      editor.on('create', handleCreate);
-
-      return () => {
-        editor.off('create', handleCreate);
-      };
-    };
+        return () => {
+          editorInstance.off('create', handleCreate);
+        };
+      },
+      [setPrevEditorHeight]
+    );
 
     useEffect(() => {
       if (editor) {
         return keepScrollPositionOnEditorReset(editor);
       }
-    }, [editor]);
+    }, [editor, keepScrollPositionOnEditorReset]);
 
-    const handleFocus = (event: React.FocusEvent<HTMLDivElement>) => {
-      setHasFocus(true);
-      onFocus?.(event as React.FocusEvent<HTMLInputElement>);
-    };
+    const handleBlurWithScrollPosition = useCallback(
+      (event: React.FocusEvent<HTMLDivElement>) => {
+        if (containerRef.current?.contains(event.relatedTarget)) {
+          return;
+        }
+        setPrevEditorHeight(editor?.view.dom.clientHeight ?? 0);
+        handleBlur(event);
+      },
+      [containerRef, editor, setPrevEditorHeight, handleBlur]
+    );
 
-    const handleBlur = (event: React.FocusEvent<HTMLDivElement>) => {
-      if (containerRef.current?.contains(event.relatedTarget)) {
-        return;
-      }
-      setPrevEditorHeight(editor?.view.dom.clientHeight ?? 0);
-      setHasFocus(false);
-      onBlur?.(event as React.FocusEvent<HTMLInputElement>);
-    };
-
-    const handleDialogOpen = useCallback(() => setIsControlsDialogOpen(true), [setIsControlsDialogOpen]);
-    const handleDialogClose = useCallback(() => setIsControlsDialogOpen(false), [setIsControlsDialogOpen]);
+    if (!editor) {
+      return null;
+    }
 
     return (
-      <Box ref={containerRef} width="100%" onFocus={handleFocus} onBlur={handleBlur}>
+      <Box ref={containerRef} width="100%" onFocus={handleFocus} onBlur={handleBlurWithScrollPosition}>
         <MarkdownInputControls
           ref={toolbarRef}
           editor={editor}
